Port to 3.10/3.11
