+++ apps/ask_demod
@@ -29,7 +29,7 @@ from gnuradio import gr
 from gnuradio import filter
 from gnuradio.filter import firdes
 
-import tpms
+import gnuradio.tpms as tpms
 
 from tpms.source import source_file
 from tpms.ask import ask_channel_filter
