+++ docs/README.tpms
@@ -2,7 +2,7 @@ This is the tpms-write-a-block package meant as a guide to building
 out-of-tree packages. To use the tpms blocks, the Python namespaces
 is in 'tpms', which is imported as:
 
-    import tpms
+    import gnuradio.tpms as tpms
 
 See the Doxygen documentation for details about the blocks available
 in this package. A quick listing of the details can be found in Python
