+++ grc/tpms_ask_env.block.yml
@@ -1,0 +1,29 @@
+# auto-generated by grc.converter
+
+id: tpms_ask_env
+label: Ask env
+category: '[TPMS]'
+
+parameters:
+-   id: alpha
+    label: Alpha
+    dtype: float
+
+inputs:
+-   domain: stream
+    dtype: float
+
+outputs:
+-   domain: stream
+    dtype: float
+
+templates:
+    imports: import gnuradio.tpms as tpms
+    make: |-
+        tpms.ask_env(
+            alpha=${alpha}
+          )
+    callbacks:
+    - set_alpha(${alpha})
+
+file_format: 1
