+++ grc/tpms_burst_detector.block.yml
@@ -1,0 +1,24 @@
+# auto-generated by grc.converter
+
+id: tpms_burst_detector
+label: burst_detector
+category: '[TPMS]'
+
+parameters:
+-   id: '...'
+    label: '...'
+    dtype: '...'
+
+inputs:
+-   domain: stream
+    dtype: ''
+
+outputs:
+-   domain: stream
+    dtype: ''
+
+templates:
+    imports: import gnuradio.tpms as tpms
+    make: tpms.burst_detector()
+
+file_format: 1
