+++ grc/tpms_fixed_length_frame_sink.block.yml
@@ -1,0 +1,24 @@
+# auto-generated by grc.converter
+
+id: tpms_fixed_length_frame_sink
+label: fixed_length_frame_sink
+category: '[TPMS]'
+
+parameters:
+-   id: '...'
+    label: '...'
+    dtype: '...'
+
+inputs:
+-   domain: stream
+    dtype: ''
+
+outputs:
+-   domain: stream
+    dtype: ''
+
+templates:
+    imports: import gnuradio.tpms as tpms
+    make: tpms.fixed_length_frame_sink(${frame_length}, ${target_queue})
+
+file_format: 1
