+++ include/gnuradio/tpms/ask_env.h
@@ -18,39 +18,38 @@
  * Boston, MA 02110-1301, USA.
  */
 
-#ifndef INCLUDED_TPMS_FIXED_LENGTH_FRAME_SINK_H
-#define INCLUDED_TPMS_FIXED_LENGTH_FRAME_SINK_H
+#ifndef INCLUDED_TPMS_ASK_ENV_H
+#define INCLUDED_TPMS_ASK_ENV_H
 
-#include <tpms/api.h>
+#include <gnuradio/tpms/api.h>
 #include <gnuradio/sync_block.h>
-#include <pmt/pmt.h>
 
 namespace gr {
   namespace tpms {
 
     /*!
-     * \brief <+description of block+>
+     * \brief Tire Pressure Monitoring System
      * \ingroup tpms
      *
      */
-    class TPMS_API fixed_length_frame_sink : virtual public gr::sync_block
+    class TPMS_API ask_env : virtual public gr::sync_block
     {
      public:
-      typedef std::shared_ptr<fixed_length_frame_sink> sptr;
+      typedef std::shared_ptr<ask_env> sptr;
 
       /*!
-       * \brief Return a shared_ptr to a new instance of tpms::fixed_length_frame_sink.
-       *
-       * To avoid accidental use of raw pointers, tpms::fixed_length_frame_sink's
-       * constructor is in a private implementation
-       * class. tpms::fixed_length_frame_sink::make is the public interface for
-       * creating new instances.
+       * \class ask_env
+       * \brief Bipolar envelope detection and scaling to +/- 1.0.
+       * \param alpha Multiplicative ate at which envelope peaks decay.
        */
-      static sptr make(int frame_length, pmt::pmt_t attributes);
+      static sptr make(float alpha);
+
+      virtual void set_alpha(float var) = 0;
+      virtual float alpha() = 0;
     };
 
   } // namespace tpms
 } // namespace gr
 
-#endif /* INCLUDED_TPMS_FIXED_LENGTH_FRAME_SINK_H */
+#endif /* INCLUDED_TPMS_ASK_ENV_H */
