+++ lib/ask_env_impl.h
@@ -21,7 +21,7 @@
 #ifndef INCLUDED_TPMS_ASK_ENV_IMPL_H
 #define INCLUDED_TPMS_ASK_ENV_IMPL_H
 
-#include <tpms/ask_env.h>
+#include <gnuradio/tpms/ask_env.h>
 
 namespace gr {
   namespace tpms {
