+++ python/tpms/bindings/ask_env_python.cc
@@ -14,7 +14,7 @@
 /* BINDTOOL_GEN_AUTOMATIC(0)                                                       */
 /* BINDTOOL_USE_PYGCCXML(0)                                                        */
 /* BINDTOOL_HEADER_FILE(ask_env.h)                                        */
-/* BINDTOOL_HEADER_FILE_HASH(2062474242a93b5a49a5242ae7ee0d7d)                     */
+/* BINDTOOL_HEADER_FILE_HASH(0021df37e7e929403d8392c5b9899e76)                     */
 /***********************************************************************************/
 
 #include <pybind11/complex.h>
@@ -23,7 +23,7 @@
 
 namespace py = pybind11;
 
-#include <tpms/ask_env.h>
+#include <gnuradio/tpms/ask_env.h>
 // pydoc.h is automatically generated in the build directory
 #include <ask_env_pydoc.h>
